Better test file organization
