+++ tests/unit/fusiontables-node.js
@@ -6,7 +6,8 @@ define(function (require) {
         assert = require('intern/chai!assert'),
         fixtures = require('tests/support/fixtures');
 
-    var FusionTables = require('fusiontables');
+    // Load FusionTables using Dojo's Node.js module loader
+    var FusionTables = require('intern/dojo/node!../../../../../src/fusiontables');
 
     function setupFT () {
         return new FusionTables({
@@ -15,13 +16,16 @@ define(function (require) {
         });
     }
 
-
-    // ~ Tests for browser request handlers (XMLHttpRequest and JSONP) ~ //
+    // ~ Tests for Node.js/io.js request handlers ~ //
 
     registerSuite({
-        name: 'FusionTables.prototype._json_request',
+        name: 'FusionTables.prototype._node_request',
 
         'fires success function with JSON data': function () {
+            if (typeof window !== 'undefined') {
+                this.skip('Node-only test');
+            }
+
             var dfd = this.async(5000);
 
             var ft = setupFT();
@@ -31,13 +35,17 @@ define(function (require) {
                 assert.strictEqual(data.hello, "world");
             });
 
-            ft._json_request(
+            ft._node_request(
                 'http://www.mocky.io/v2/5681b3ea120000952293a28b',
                 success
             );
         },
 
         'fires error function on network error': function () {
+            if (typeof window !== 'undefined') {
+                this.skip('Node-only test');
+            }
+
             var dfd = this.async(5000);
 
             var ft = setupFT();
@@ -46,7 +54,7 @@ define(function (require) {
                 assert.instanceOf(e, Error);
             });
 
-            ft._json_request(
+            ft._node_request(
                 'http://www.a5681b3ea120000952293a28b.com/',
                 function() {},
                 error
@@ -54,6 +62,10 @@ define(function (require) {
         },
 
         'fires error function on HTTP error': function () {
+            if (typeof window !== 'undefined') {
+                this.skip('Node-only test');
+            }
+
             var dfd = this.async(5000);
 
             var ft = setupFT();
@@ -62,44 +74,8 @@ define(function (require) {
                 assert.instanceOf(e, Error);
             });
 
-            ft._json_request(
+            ft._node_request(
                 'http://www.example.com/404',
-                function() {},
-                error
-            );
-        }
-    });
-
-    registerSuite({
-        name: 'FusionTables.prototype._jsonp_request',
-
-        'fires success function with JSON data': function () {
-            var dfd = this.async(5000);
-
-            var ft = setupFT();
-
-            var success = dfd.callback(function (data) {
-                assert.instanceOf(data, Object);
-                assert.deepEqual(data, { hello : "world" });
-            });
-
-            ft._jsonp_request(
-                'http://www.mocky.io/v2/5681b3ea120000952293a28b',
-                success
-            );
-        },
-
-        'fires error function on network/HTTP error': function () {
-            var dfd = this.async(5000);
-
-            var ft = setupFT();
-
-            var error = dfd.callback(function (e) {
-                assert.instanceOf(e, Error);
-            });
-
-            ft._jsonp_request(
-                'http://www.a5681b3ea120000952293a28b.com/',
                 function() {},
                 error
             );
